refactor(mcp): split server.rs into focused submodules
