+++ src-tauri/src/mcp/server.rs
@@ -1,136 +1,22 @@
-//! DataZen MCP Server handler — tools + resources + prompts.
+//! DataZen MCP Server — orchestration entry (struct, core helpers, tests).
 
-use crate::ai::budget;
-use crate::ai::prompt_resolver;
 use crate::commands::AppState;
 use crate::mcp::allowlist;
-use crate::mcp::permission::{self, McpPermissionMode};
+use crate::mcp::permission::McpPermissionMode;
 use crate::mcp::tool_help;
-use datazen_driver_api::PromptScenario;
-use rmcp::handler::server::wrapper::Parameters;
-use rmcp::model::*;
-use rmcp::service::RequestContext;
-use rmcp::{
-    prompt, prompt_handler, prompt_router, tool, tool_handler, tool_router, ErrorData as McpError,
-    RoleServer, ServerHandler,
-};
-use schemars::JsonSchema;
-use serde::Deserialize;
-use std::collections::{HashMap, HashSet};
+use rmcp::ErrorData as McpError;
+use std::collections::HashSet;
 use std::sync::Arc;
 
-// ─── Tool Input Types ───
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct QueryInput {
-    /// Persistent connection id (from list_connections)
-    pub connection_id: String,
-    /// SQL query to execute
-    pub sql: String,
-    /// Maximum rows to return (default: 100, max: 50000)
-    pub limit: Option<u32>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct ListTablesInput {
-    /// Persistent connection id (from list_connections)
-    pub connection_id: String,
-    /// Optional database name
-    pub database: Option<String>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct SearchTablesInput {
-    /// Persistent connection id (from list_connections)
-    pub connection_id: String,
-    /// Optional database name
-    pub database: Option<String>,
-    /// Search keyword to match against table names (case-insensitive)
-    pub pattern: String,
-    /// Max results to return (default: 20)
-    pub limit: Option<u64>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct GetSchemaInput {
-    /// Persistent connection id (from list_connections)
-    pub connection_id: String,
-    /// Table name
-    pub table: String,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct ExplainQueryInput {
-    /// Persistent connection id (from list_connections)
-    pub connection_id: String,
-    /// SQL query to analyze
-    pub sql: String,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct DescribeTableInput {
-    /// Persistent connection id (from list_connections)
-    pub connection_id: String,
-    /// Table name
-    pub table: String,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct ListDatabasesInput {
-    /// Persistent connection id (from list_connections)
-    pub connection_id: String,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct RunWorkflowInput {
-    /// Workflow ID to execute
-    pub workflow_id: String,
-    /// Input variables for the workflow (JSON object)
-    #[serde(default)]
-    pub variables: serde_json::Value,
-    /// Optional persistent connection id (some workflows require a database connection)
-    pub connection_id: Option<String>,
-}
-
-// ─── Prompt Argument Types ───
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct Nl2SqlArgs {
-    /// Persistent connection id for schema context
-    pub connection_id: String,
-    /// Natural language description of the query
-    pub question: String,
-    /// Optional database name
-    pub database: Option<String>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct DiagnoseErrorArgs {
-    /// Persistent connection id
-    pub connection_id: String,
-    /// The SQL that caused the error
-    pub sql: String,
-    /// The error message
-    pub error: String,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct ExplainPlanArgs {
-    /// Persistent connection id
-    pub connection_id: String,
-    /// The SQL query to explain
-    pub sql: String,
-}
-
-// ─── MCP Server ───
+pub use super::types::*;
 
 #[derive(Clone)]
 pub struct DataZenMcpServer {
-    app_state: Arc<AppState>,
-    disabled_tools: HashSet<String>,
-    permission_mode: McpPermissionMode,
+    pub(crate) app_state: Arc<AppState>,
+    pub(crate) disabled_tools: HashSet<String>,
+    pub(crate) permission_mode: McpPermissionMode,
     /// Empty = all saved connections are exposed to MCP.
-    allowed_connection_ids: Vec<String>,
+    pub(crate) allowed_connection_ids: Vec<String>,
 }
 
 pub const MCP_ALL_TOOLS: &[&str] = &[
@@ -171,20 +57,20 @@ impl DataZenMcpServer {
         self
     }
 
-    fn map_err(e: String) -> McpError {
+    pub(crate) fn map_err(e: String) -> McpError {
         McpError::internal_error(e, None)
     }
 
-    fn ensure_allowed(&self, tool_name: &str, connection_id: &str) -> Result<(), McpError> {
+    pub(crate) fn ensure_allowed(&self, tool_name: &str, connection_id: &str) -> Result<(), McpError> {
         allowlist::ensure_connection_allowed(connection_id, &self.allowed_connection_ids)
             .map_err(|e| tool_help::tool_error(tool_name, &e))
     }
 
-    fn tool_is_registered(&self, name: &str) -> bool {
+    pub(crate) fn tool_is_registered(&self, name: &str) -> bool {
         Self::tool_router().get(name).is_some()
     }
 
-    async fn resolve_connection(
+    pub(crate) async fn resolve_connection(
         &self,
         id: &str,
     ) -> Result<
@@ -205,699 +91,11 @@ impl DataZenMcpServer {
     }
 }
 
-/// Human-readable table description for MCP `describe_table` tool output.
-pub(crate) fn format_table_description(
-    table: &str,
-    schema: &datazen_driver_api::TableSchema,
-) -> String {
-    let mut desc = format!("Table: {table}\n\nColumns:\n");
-    for col in &schema.columns {
-        desc.push_str(&format!(
-            "  - {} {} {}{}\n",
-            col.name,
-            col.data_type,
-            if col.is_primary_key { "PK " } else { "" },
-            if col.nullable { "" } else { "NOT NULL " },
-        ));
-    }
-
-    if !schema.primary_keys.is_empty() {
-        desc.push_str(&format!(
-            "\nPrimary Key: ({})\n",
-            schema.primary_keys.join(", ")
-        ));
-    }
-
-    if !schema.indexes.is_empty() {
-        desc.push_str("\nIndexes:\n");
-        for idx in &schema.indexes {
-            desc.push_str(&format!(
-                "  - {} ({}) {}\n",
-                idx.name,
-                idx.columns.join(", "),
-                if idx.is_unique { "UNIQUE" } else { "" }
-            ));
-        }
-    }
-
-    if !schema.foreign_keys.is_empty() {
-        desc.push_str("\nForeign Keys:\n");
-        for fk in &schema.foreign_keys {
-            desc.push_str(&format!(
-                "  - {} ({}) → {}.{}\n",
-                fk.name,
-                fk.columns.join(", "),
-                fk.referenced_table,
-                fk.referenced_columns.join(", ")
-            ));
-        }
-    }
-
-    desc
-}
-
-// ─── Tools ───
-
-#[tool_router]
-impl DataZenMcpServer {
-    #[tool(
-        description = "List all configured database connections. Returns connection IDs, names, database types, and hosts."
-    )]
-    async fn list_connections(&self) -> Result<String, McpError> {
-        let connections = self.app_state.store.get_connections().await;
-        let result: Vec<serde_json::Value> = connections
-            .iter()
-            .filter(|c| allowlist::is_connection_allowed(&c.id, &self.allowed_connection_ids))
-            .map(|c| {
-                serde_json::json!({
-                    "id": c.id,
-                    "name": c.name,
-                    "databaseType": format!("{:?}", c.database_type),
-                    "host": c.host,
-                    "database": c.database,
-                })
-            })
-            .collect();
-        serde_json::to_string_pretty(&result).map_err(|e| Self::map_err(format!("Error: {e}")))
-    }
-
-    #[tool(description = "List all databases on a connected server.")]
-    async fn list_databases(
-        &self,
-        Parameters(input): Parameters<ListDatabasesInput>,
-    ) -> Result<String, McpError> {
-        self.ensure_allowed("list_databases", &input.connection_id)?;
-        crate::services::db_tools::list_databases(
-            &self.app_state.connection_manager,
-            &input.connection_id,
-        )
-        .await
-        .map_err(Self::map_err)
-    }
-
-    #[tool(
-        description = "List all tables in a database. Returns table names with types and row counts."
-    )]
-    async fn list_tables(
-        &self,
-        Parameters(input): Parameters<ListTablesInput>,
-    ) -> Result<String, McpError> {
-        self.ensure_allowed("list_tables", &input.connection_id)?;
-        let db = input.database.as_deref().unwrap_or("");
-        crate::services::db_tools::list_tables(
-            &self.app_state.connection_manager,
-            &input.connection_id,
-            db,
-        )
-        .await
-        .map_err(Self::map_err)
-    }
-
-    #[tool(
-        description = "Search for tables by name pattern (case-insensitive substring match). Use this instead of list_tables when the database has many tables."
-    )]
-    async fn search_tables(
-        &self,
-        Parameters(input): Parameters<SearchTablesInput>,
-    ) -> Result<String, McpError> {
-        self.ensure_allowed("search_tables", &input.connection_id)?;
-        let db = input.database.as_deref().unwrap_or("");
-        let limit = input.limit.unwrap_or(20) as usize;
-        crate::services::db_tools::search_tables(
-            &self.app_state.connection_manager,
-            &input.connection_id,
-            db,
-            &input.pattern,
-            limit,
-        )
-        .await
-        .map_err(Self::map_err)
-    }
-
-    #[tool(
-        description = "Execute a SQL query on a connected database. Returns results as JSON. Use list_connections first to get valid connection IDs. Default row limit is 100 (max 50000)."
-    )]
-    async fn query(&self, Parameters(input): Parameters<QueryInput>) -> Result<String, McpError> {
-        self.ensure_allowed("query", &input.connection_id)?;
-        permission::check_sql_allowed(&input.sql, self.permission_mode)
-            .map_err(|e| tool_help::tool_error("query", &e))?;
-        crate::services::db_tools::query(
-            &self.app_state.connection_manager,
-            &input.connection_id,
-            &input.sql,
-            input.limit,
-            Some(self.permission_mode),
-        )
-        .await
-        .map_err(Self::map_err)
-    }
-
-    #[tool(
-        description = "Get detailed schema of a table including columns, types, primary keys, foreign keys, and indexes."
-    )]
-    async fn get_schema(
-        &self,
-        Parameters(input): Parameters<GetSchemaInput>,
-    ) -> Result<String, McpError> {
-        self.ensure_allowed("get_schema", &input.connection_id)?;
-        let tables = vec![input.table.clone()];
-        crate::services::db_tools::get_table_schema(
-            &self.app_state.connection_manager,
-            &input.connection_id,
-            &tables,
-        )
-        .await
-        .map_err(Self::map_err)
-    }
-
-    #[tool(
-        description = "Get the execution plan (EXPLAIN) for a SQL query. Useful for performance analysis."
-    )]
-    async fn explain_query(
-        &self,
-        Parameters(input): Parameters<ExplainQueryInput>,
-    ) -> Result<String, McpError> {
-        self.ensure_allowed("explain_query", &input.connection_id)?;
-        crate::services::db_tools::explain_query(
-            &self.app_state.connection_manager,
-            &input.connection_id,
-            &input.sql,
-        )
-        .await
-        .map_err(Self::map_err)
-    }
-
-    #[tool(
-        description = "Get a human-readable description of a table including columns, types, constraints, and indexes."
-    )]
-    async fn describe_table(
-        &self,
-        Parameters(input): Parameters<DescribeTableInput>,
-    ) -> Result<String, McpError> {
-        self.ensure_allowed("describe_table", &input.connection_id)?;
-        let schema = crate::services::db_tools::get_single_table_schema(
-            &self.app_state.connection_manager,
-            &input.connection_id,
-            &input.table,
-        )
-        .await
-        .map_err(Self::map_err)?;
-
-        Ok(format_table_description(&input.table, &schema))
-    }
-
-    #[tool(
-        description = "List all available user-defined workflows. Workflows are reusable AI automations combining prompts and database operations."
-    )]
-    async fn list_workflows(&self) -> Result<String, McpError> {
-        let workflows = self.app_state.workflow_registry.list().await;
-        serde_json::to_string_pretty(&workflows)
-            .map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    #[tool(
-        description = "Execute a user-defined workflow by ID. Workflows are reusable automations combining prompts and database operations. Use list_workflows to see available workflows."
-    )]
-    async fn run_workflow(
-        &self,
-        Parameters(input): Parameters<RunWorkflowInput>,
-    ) -> Result<String, McpError> {
-        let workflow = self
-            .app_state
-            .workflow_registry
-            .get(&input.workflow_id)
-            .await
-            .ok_or_else(|| {
-                tool_help::tool_error(
-                    "run_workflow",
-                    &format!("Workflow '{}' not found", input.workflow_id),
-                )
-            })?;
-
-        let result = crate::workflow::WorkflowExecutor::execute_with_options(
-            &workflow,
-            &self.app_state,
-            input.connection_id.as_deref(),
-            &input.variables,
-            crate::workflow::WorkflowExecuteOptions {
-                permission_mode: Some(self.permission_mode),
-                query_row_limit: Some(crate::workflow::WORKFLOW_QUERY_ROW_LIMIT),
-            },
-        )
-        .await
-        .map_err(|e| McpError::internal_error(e, None))?;
-
-        Ok(result.final_output)
-    }
-}
-
-// ─── Prompts ───
-
-#[prompt_router]
-impl DataZenMcpServer {
-    #[prompt(
-        name = "nl2sql",
-        description = "Convert natural language to SQL based on the database schema"
-    )]
-    async fn nl2sql_prompt(
-        &self,
-        Parameters(args): Parameters<Nl2SqlArgs>,
-    ) -> Result<GetPromptResult, McpError> {
-        let (conn_id, driver, _handle) = self.resolve_connection(&args.connection_id).await?;
-        let lang = self.app_state.store.get_settings().await.language;
-        let db_type = driver.driver_type();
-        let db = args.database.as_deref().unwrap_or("");
-
-        let context = self
-            .app_state
-            .schema_context_builder
-            .build_sql_context(&conn_id, db, None, &[], budget::FALLBACK_DDL)
-            .await
-            .map_err(|e| McpError::internal_error(e.to_string(), None))?;
-
-        let mut vars = HashMap::new();
-        vars.insert("db_type", db_type.as_str());
-        vars.insert("version", "");
-        vars.insert("schema", context.schema_ddl.as_str());
-        vars.insert("recent", "");
-        let system = self
-            .app_state
-            .prompt_resolver
-            .resolve(PromptScenario::Nl2Sql, Some(driver.as_ref()), &lang)
-            .await;
-        let system_content = prompt_resolver::render_template(&system, &vars);
-
-        Ok(GetPromptResult::new(vec![
-            PromptMessage::new_text(Role::User, system_content),
-            PromptMessage::new_text(Role::User, args.question.clone()),
-        ])
-        .with_description("Natural language to SQL conversion with schema context"))
-    }
-
-    #[prompt(
-        name = "diagnose_error",
-        description = "Diagnose a SQL error and suggest fixes"
-    )]
-    async fn diagnose_error_prompt(
-        &self,
-        Parameters(args): Parameters<DiagnoseErrorArgs>,
-    ) -> Result<GetPromptResult, McpError> {
-        let (_conn_id, driver, _handle) = self.resolve_connection(&args.connection_id).await?;
-        let lang = self.app_state.store.get_settings().await.language;
-        let db_type = driver.driver_type();
-
-        let mut vars = HashMap::new();
-        vars.insert("db_type", db_type.as_str());
-        vars.insert("version", "");
-        vars.insert("schema", "");
-        vars.insert("recent", "");
-        let system = self
-            .app_state
-            .prompt_resolver
-            .resolve(PromptScenario::Diagnose, Some(driver.as_ref()), &lang)
-            .await;
-        let system_content = prompt_resolver::render_template(&system, &vars);
-
-        Ok(GetPromptResult::new(vec![
-            PromptMessage::new_text(Role::User, system_content),
-            PromptMessage::new_text(
-                Role::User,
-                format!("SQL:\n```\n{}\n```\n\nError:\n{}", args.sql, args.error),
-            ),
-        ])
-        .with_description("SQL error diagnosis with fix suggestions"))
-    }
-
-    #[prompt(
-        name = "explain_plan",
-        description = "Analyze a query execution plan and suggest optimizations"
-    )]
-    async fn explain_plan_prompt(
-        &self,
-        Parameters(args): Parameters<ExplainPlanArgs>,
-    ) -> Result<GetPromptResult, McpError> {
-        let (_conn_id, driver, handle) = self.resolve_connection(&args.connection_id).await?;
-        let lang = self.app_state.store.get_settings().await.language;
-        let db_type = driver.driver_type();
-
-        let explain_result = driver
-            .explain(&handle, &args.sql)
-            .await
-            .map_err(|e| McpError::internal_error(e.to_string(), None))?;
-
-        let explain_text = serde_json::to_string_pretty(&explain_result)
-            .unwrap_or_else(|_| "Failed to serialize EXPLAIN output".to_string());
-
-        let mut vars = HashMap::new();
-        vars.insert("db_type", db_type.as_str());
-        vars.insert("version", "");
-        vars.insert("schema", "");
-        vars.insert("recent", "");
-        let system = self
-            .app_state
-            .prompt_resolver
-            .resolve(
-                PromptScenario::ExplainAnalysis,
-                Some(driver.as_ref()),
-                &lang,
-            )
-            .await;
-        let system_content = prompt_resolver::render_template(&system, &vars);
-
-        Ok(GetPromptResult::new(vec![
-            PromptMessage::new_text(Role::User, system_content),
-            PromptMessage::new_text(
-                Role::User,
-                format!(
-                    "SQL:\n```\n{}\n```\n\nEXPLAIN output:\n```\n{}\n```",
-                    args.sql, explain_text
-                ),
-            ),
-        ])
-        .with_description("Query execution plan analysis"))
-    }
-}
-
-impl DataZenMcpServer {
-    pub(crate) async fn read_resource_inner(
-        &self,
-        uri: &str,
-    ) -> Result<ReadResourceResult, McpError> {
-        if uri == "datazen://connections" {
-            let connections = self.app_state.store.get_connections().await;
-            let result: Vec<serde_json::Value> = connections
-                .iter()
-                .filter(|c| allowlist::is_connection_allowed(&c.id, &self.allowed_connection_ids))
-                .map(|c| {
-                    serde_json::json!({
-                        "id": c.id,
-                        "name": c.name,
-                        "databaseType": format!("{:?}", c.database_type),
-                        "host": c.host,
-                        "database": c.database,
-                    })
-                })
-                .collect();
-            let json = serde_json::to_string_pretty(&result)
-                .map_err(|e| McpError::internal_error(e.to_string(), None))?;
-            return Ok(ReadResourceResult::new(vec![ResourceContents::text(
-                json, uri,
-            )]));
-        }
-
-        if uri == "datazen://workflows" {
-            let workflows = self.app_state.workflow_registry.list().await;
-            let json = serde_json::to_string_pretty(&workflows)
-                .map_err(|e| McpError::internal_error(e.to_string(), None))?;
-            return Ok(ReadResourceResult::new(vec![ResourceContents::text(
-                json, uri,
-            )]));
-        }
-
-        if uri == "datazen://query-history" {
-            if !permission::query_history_content_allowed(self.permission_mode) {
-                return Ok(ReadResourceResult::new(vec![ResourceContents::text(
-                    "[]", uri,
-                )]));
-            }
-            let history = self
-                .app_state
-                .store
-                .get_query_history(50, None, None, None)
-                .await;
-            let filtered: Vec<_> = history
-                .into_iter()
-                .filter(|entry| {
-                    allowlist::is_connection_allowed(
-                        &entry.connection_id,
-                        &self.allowed_connection_ids,
-                    )
-                })
-                .collect();
-            let json = serde_json::to_string_pretty(&filtered)
-                .map_err(|e| McpError::internal_error(e.to_string(), None))?;
-            return Ok(ReadResourceResult::new(vec![ResourceContents::text(
-                json, uri,
-            )]));
-        }
-
-        if let Some(rest) = uri.strip_prefix("datazen://schema/") {
-            let decoded = urlencoding::decode(rest).map_err(|e| {
-                McpError::invalid_params(format!("Invalid URI encoding: {e}"), None)
-            })?;
-            let parts: Vec<&str> = decoded.splitn(2, '/').collect();
-            let connection_id = parts
-                .first()
-                .ok_or_else(|| McpError::invalid_params("Missing connection ID in URI", None))?;
-            let db = parts.get(1).copied().unwrap_or("");
-
-            let (runtime_id, _driver, _handle) = self.resolve_connection(connection_id).await?;
-
-            let context = self
-                .app_state
-                .schema_context_builder
-                .build_sql_context(&runtime_id, db, None, &[], budget::MCP_RESOURCE)
-                .await
-                .map_err(|e| McpError::internal_error(e.to_string(), None))?;
-
-            return Ok(ReadResourceResult::new(vec![ResourceContents::text(
-                context.schema_ddl,
-                uri,
-            )]));
-        }
-
-        Err(McpError::resource_not_found(
-            format!("Unknown resource: {uri}"),
-            None,
-        ))
-    }
-
-    #[cfg(test)]
-    pub(crate) async fn call_tool_inner(
-        &self,
-        name: &str,
-        arguments: Option<serde_json::Value>,
-    ) -> Result<CallToolResult, McpError> {
-        let arg_map = arguments.as_ref().and_then(|v| v.as_object());
-        permission::check_tool_call(name, self.permission_mode, &self.disabled_tools, arg_map)
-            .map_err(|e| tool_help::tool_error(name, &e))?;
-
-        let text = match name {
-            "list_connections" => self.list_connections().await?,
-            "list_databases" => {
-                let connection_id = arguments
-                    .as_ref()
-                    .and_then(|a| a.get("connection_id"))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-                self.list_databases(rmcp::handler::server::wrapper::Parameters(
-                    ListDatabasesInput {
-                        connection_id: connection_id.into(),
-                    },
-                ))
-                .await?
-            }
-            "list_tables" => {
-                let connection_id = arguments
-                    .as_ref()
-                    .and_then(|a| a.get("connection_id"))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-                let database = arguments
-                    .as_ref()
-                    .and_then(|a| a.get("database"))
-                    .and_then(|v| v.as_str())
-                    .map(String::from);
-                self.list_tables(rmcp::handler::server::wrapper::Parameters(
-                    ListTablesInput {
-                        connection_id: connection_id.into(),
-                        database,
-                    },
-                ))
-                .await?
-            }
-            "search_tables" => {
-                let connection_id = arguments
-                    .as_ref()
-                    .and_then(|a| a.get("connection_id"))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-                let database = arguments
-                    .as_ref()
-                    .and_then(|a| a.get("database"))
-                    .and_then(|v| v.as_str())
-                    .map(String::from);
-                let pattern = arguments
-                    .as_ref()
-                    .and_then(|a| a.get("pattern"))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-                let limit = arguments
-                    .as_ref()
-                    .and_then(|a| a.get("limit"))
-                    .and_then(|v| v.as_u64());
-                self.search_tables(rmcp::handler::server::wrapper::Parameters(
-                    SearchTablesInput {
-                        connection_id: connection_id.into(),
-                        database,
-                        pattern: pattern.into(),
-                        limit,
-                    },
-                ))
-                .await?
-            }
-            "query" => {
-                let connection_id = arguments
-                    .as_ref()
-                    .and_then(|a| a.get("connection_id"))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-                let sql = arguments
-                    .as_ref()
-                    .and_then(|a| a.get("sql"))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-                let limit = arguments
-                    .as_ref()
-                    .and_then(|a| a.get("limit"))
-                    .and_then(|v| v.as_u64())
-                    .map(|n| n as u32);
-                self.query(rmcp::handler::server::wrapper::Parameters(QueryInput {
-                    connection_id: connection_id.into(),
-                    sql: sql.into(),
-                    limit,
-                }))
-                .await?
-            }
-            other => return Err(tool_help::unknown_tool_error(other)),
-        };
-
-        Ok(CallToolResult::success(vec![ContentBlock::text(text)]))
-    }
-
-    pub(crate) fn list_tools_inner(&self) -> Vec<Tool> {
-        let mut router = Self::tool_router();
-        for name in MCP_ALL_TOOLS {
-            if !permission::is_tool_listed(name, self.permission_mode, &self.disabled_tools) {
-                router.disable_route(name.to_string());
-            }
-        }
-        router.list_all()
-    }
-
-    pub(crate) fn list_resources_inner(&self) -> Vec<Resource> {
-        [
-            ("datazen://connections", "Database Connections"),
-            ("datazen://query-history", "Query History"),
-            ("datazen://workflows", "Available Workflows"),
-        ]
-        .into_iter()
-        .filter(|(uri, _)| permission::is_resource_listed(uri, self.permission_mode))
-        .map(|(uri, name)| Resource::new(uri, name))
-        .collect()
-    }
-}
-
-// ─── ServerHandler (resources + capabilities) ───
-
-#[tool_handler]
-#[prompt_handler]
-impl ServerHandler for DataZenMcpServer {
-    async fn list_tools(
-        &self,
-        _request: Option<PaginatedRequestParams>,
-        _context: RequestContext<RoleServer>,
-    ) -> Result<ListToolsResult, McpError> {
-        Ok(ListToolsResult {
-            tools: self.list_tools_inner(),
-            next_cursor: None,
-            meta: None,
-        })
-    }
-
-    async fn call_tool(
-        &self,
-        request: CallToolRequestParams,
-        context: RequestContext<RoleServer>,
-    ) -> Result<CallToolResult, McpError> {
-        let tool_name = request.name.to_string();
-
-        if !self.tool_is_registered(&tool_name) {
-            return Err(tool_help::unknown_tool_error(&tool_name));
-        }
-
-        permission::check_tool_call(
-            &tool_name,
-            self.permission_mode,
-            &self.disabled_tools,
-            request.arguments.as_ref(),
-        )
-        .map_err(|e| tool_help::tool_error(&tool_name, &e))?;
-
-        if let Some(args) = request.arguments.as_ref() {
-            if let Some(connection_id) = args.get("connection_id").and_then(|v| v.as_str()) {
-                self.ensure_allowed(&tool_name, connection_id)?;
-            }
-        }
-
-        let tcc = rmcp::handler::server::tool::ToolCallContext::new(self, request, context);
-        Self::tool_router()
-            .call(tcc)
-            .await
-            .map_err(|e| tool_help::enrich_tool_error(&tool_name, e))
-    }
-
-    fn get_info(&self) -> ServerInfo {
-        ServerInfo::new(
-            ServerCapabilities::builder()
-                .enable_tools()
-                .enable_resources()
-                .enable_prompts()
-                .build(),
-        )
-        .with_server_info(Implementation::new("datazen", env!("CARGO_PKG_VERSION")))
-        .with_instructions("DataZen is a database management tool. Use list_connections to discover available databases, then query them with SQL.".to_string())
-    }
-
-    async fn list_resources(
-        &self,
-        _request: Option<PaginatedRequestParams>,
-        _context: RequestContext<RoleServer>,
-    ) -> Result<ListResourcesResult, McpError> {
-        Ok(ListResourcesResult {
-            resources: self.list_resources_inner(),
-            next_cursor: None,
-            meta: None,
-        })
-    }
-
-    async fn list_resource_templates(
-        &self,
-        _request: Option<PaginatedRequestParams>,
-        _context: RequestContext<RoleServer>,
-    ) -> Result<ListResourceTemplatesResult, McpError> {
-        Ok(ListResourceTemplatesResult {
-            resource_templates: vec![ResourceTemplate::new(
-                "datazen://schema/{connectionId}/{database}",
-                "Database Schema",
-            )],
-            next_cursor: None,
-            meta: None,
-        })
-    }
-
-    async fn read_resource(
-        &self,
-        request: ReadResourceRequestParams,
-        _context: RequestContext<RoleServer>,
-    ) -> Result<ReadResourceResult, McpError> {
-        self.read_resource_inner(request.uri.as_str()).await
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::mcp::tools::format_table_description;
+    use rmcp::model::ResourceContents;
 
     #[test]
     fn test_uri_encoding_decoding() {
@@ -1083,620 +281,6 @@ mod tests {
         assert!(desc.contains("users_pkey"));
     }
 
-    #[tokio::test]
-    async fn mcp_tool_handlers_with_mock_driver() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        test.save_connection("mcp-cfg").await;
-        let server = DataZenMcpServer::new(Arc::new(test.state));
-
-        let conns = server.list_connections().await.unwrap();
-        assert!(conns.contains("mcp-cfg"));
-
-        let dbs = server
-            .list_databases(rmcp::handler::server::wrapper::Parameters(
-                ListDatabasesInput {
-                    connection_id: "mcp-cfg".into(),
-                },
-            ))
-            .await
-            .unwrap();
-        assert!(dbs.contains("app"));
-
-        let tables = server
-            .list_tables(rmcp::handler::server::wrapper::Parameters(
-                ListTablesInput {
-                    connection_id: "mcp-cfg".into(),
-                    database: Some("app".into()),
-                },
-            ))
-            .await
-            .unwrap();
-        assert!(tables.contains("users"));
-
-        let schema = server
-            .get_schema(rmcp::handler::server::wrapper::Parameters(GetSchemaInput {
-                connection_id: "mcp-cfg".into(),
-                table: "users".into(),
-            }))
-            .await
-            .unwrap();
-        assert!(schema.contains("users"));
-
-        let desc = server
-            .describe_table(rmcp::handler::server::wrapper::Parameters(
-                DescribeTableInput {
-                    connection_id: "mcp-cfg".into(),
-                    table: "users".into(),
-                },
-            ))
-            .await
-            .unwrap();
-        assert!(desc.contains("Table: users"));
-
-        let explain = server
-            .explain_query(rmcp::handler::server::wrapper::Parameters(
-                ExplainQueryInput {
-                    connection_id: "mcp-cfg".into(),
-                    sql: "SELECT 1".into(),
-                },
-            ))
-            .await
-            .unwrap();
-        assert!(!explain.is_empty());
-
-        let query_out = server
-            .query(rmcp::handler::server::wrapper::Parameters(QueryInput {
-                connection_id: "mcp-cfg".into(),
-                sql: "SELECT 1".into(),
-                limit: Some(10),
-            }))
-            .await
-            .unwrap();
-        assert!(query_out.contains("alice") || query_out.contains("1"));
-
-        let workflows = server.list_workflows().await.unwrap();
-        assert!(
-            workflows.contains("builtin-hello-query") || workflows.contains("[]"),
-            "expected builtin workflows or an empty list, got: {workflows}"
-        );
-
-        let search_result = server
-            .search_tables(rmcp::handler::server::wrapper::Parameters(
-                SearchTablesInput {
-                    connection_id: "mcp-cfg".into(),
-                    database: Some("app".into()),
-                    pattern: "user".into(),
-                    limit: Some(10),
-                },
-            ))
-            .await
-            .unwrap();
-        assert!(search_result.contains("users"));
-        assert!(search_result.contains("totalMatches"));
-
-        let search_no_match = server
-            .search_tables(rmcp::handler::server::wrapper::Parameters(
-                SearchTablesInput {
-                    connection_id: "mcp-cfg".into(),
-                    database: Some("app".into()),
-                    pattern: "zzz_nonexistent".into(),
-                    limit: None,
-                },
-            ))
-            .await
-            .unwrap();
-        let parsed: serde_json::Value = serde_json::from_str(&search_no_match).unwrap();
-        assert_eq!(parsed["totalMatches"].as_u64(), Some(0));
-    }
-
-    #[tokio::test]
-    async fn mcp_search_tables_via_call_tool_inner() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        test.save_connection("cti-search").await;
-        let server = DataZenMcpServer::new(Arc::new(test.state));
-
-        let result = server
-            .call_tool_inner(
-                "search_tables",
-                Some(
-                    serde_json::json!({"connection_id":"cti-search","database":"app","pattern":"user"}),
-                ),
-            )
-            .await
-            .unwrap();
-        let text = result.content[0].as_text().unwrap().text.as_str();
-        assert!(text.contains("users"));
-        assert!(text.contains("totalMatches"));
-    }
-
-    #[tokio::test]
-    async fn mcp_prompt_handlers_build_messages() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        let (config, _db_session_id) = test.save_and_connect("mcp-prompt").await;
-        let connection_id = config.id.clone();
-        let server = DataZenMcpServer::new(Arc::new(test.state));
-
-        let nl2sql = server
-            .nl2sql_prompt(rmcp::handler::server::wrapper::Parameters(Nl2SqlArgs {
-                connection_id: connection_id.clone(),
-                question: "count users".into(),
-                database: Some("app".into()),
-            }))
-            .await
-            .unwrap();
-        assert!(nl2sql.messages[0]
-            .content
-            .as_text()
-            .unwrap()
-            .text
-            .contains("Schema:"));
-        assert!(nl2sql.messages[1]
-            .content
-            .as_text()
-            .unwrap()
-            .text
-            .contains("count users"));
-
-        let diag = server
-            .diagnose_error_prompt(rmcp::handler::server::wrapper::Parameters(
-                DiagnoseErrorArgs {
-                    connection_id: connection_id.clone(),
-                    sql: "SELECT bad".into(),
-                    error: "column missing".into(),
-                },
-            ))
-            .await
-            .unwrap();
-        assert!(diag.messages[1]
-            .content
-            .as_text()
-            .unwrap()
-            .text
-            .contains("column missing"));
-
-        let plan = server
-            .explain_plan_prompt(rmcp::handler::server::wrapper::Parameters(
-                ExplainPlanArgs {
-                    connection_id,
-                    sql: "SELECT 1".into(),
-                },
-            ))
-            .await
-            .unwrap();
-        assert!(plan.messages[1]
-            .content
-            .as_text()
-            .unwrap()
-            .text
-            .contains("EXPLAIN"));
-    }
-
-    #[tokio::test]
-    async fn mcp_tool_router_lists_registered_tools() {
-        let tools = DataZenMcpServer::tool_router().list_all();
-        assert_eq!(tools.len(), MCP_ALL_TOOLS.len());
-        assert!(tools.iter().any(|t| t.name == "query"));
-    }
-
-    #[tokio::test]
-    async fn mcp_read_resource_inner_paths() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        test.save_connection("res-inner").await;
-        let (_config, _db_session_id) = test.save_and_connect("res-inner").await;
-        let server = DataZenMcpServer::new(Arc::new(test.state));
-
-        let conns = server
-            .read_resource_inner("datazen://connections")
-            .await
-            .unwrap();
-        match &conns.contents[0] {
-            ResourceContents::TextResourceContents { text, .. } => {
-                assert!(text.contains("res-inner"));
-            }
-            _ => panic!("expected text resource"),
-        }
-
-        // Seed one row so the resource contract below is exercised on real output.
-        test.store
-            .add_query_history(crate::store::QueryHistoryEntry {
-                id: "hist-1".into(),
-                connection_id: "res-inner".into(),
-                database: "app".into(),
-                schema: None,
-                sql: "SELECT 1".into(),
-                executed_at: chrono::Utc::now(),
-                execution_time_ms: 1,
-                rows_affected: None,
-                success: true,
-                error_message: None,
-            })
-            .await
-            .expect("add_query_history");
-
-        let hist = server
-            .read_resource_inner("datazen://query-history")
-            .await
-            .unwrap();
-        // External contract: history entries must serialize with `connectionId`
-        // (camelCase of `connection_id`); the legacy `configId` key is gone.
-        assert!(!hist.contents.is_empty());
-        match &hist.contents[0] {
-            ResourceContents::TextResourceContents { text, .. } => {
-                assert!(
-                    text.contains("\"connectionId\""),
-                    "query-history resource must expose connectionId, got: {text}"
-                );
-                assert!(
-                    !text.contains("configId"),
-                    "legacy configId key must not appear in resource output"
-                );
-            }
-            _ => panic!("expected text resource"),
-        }
-
-        let schema_uri = "datazen://schema/res-inner/app";
-        let schema = server.read_resource_inner(&schema_uri).await.unwrap();
-        match &schema.contents[0] {
-            ResourceContents::TextResourceContents { text, .. } => {
-                assert!(!text.is_empty());
-            }
-            _ => panic!("expected text resource"),
-        }
-
-        let err = server
-            .read_resource_inner("datazen://missing")
-            .await
-            .unwrap_err();
-        assert!(err.to_string().contains("Unknown resource"));
-    }
-
-    #[tokio::test]
-    async fn mcp_call_tool_inner_unknown_tool_includes_help() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::new().await;
-        let server = DataZenMcpServer::new(Arc::new(test.state));
-
-        let err = server
-            .call_tool_inner("missing_tool", None)
-            .await
-            .unwrap_err();
-        let msg = err.message.to_string();
-        assert!(msg.contains("Unknown or disabled tool"));
-        assert!(msg.contains("list_connections"));
-        assert!(msg.contains("Example"));
-    }
-
-    #[tokio::test]
-    async fn mcp_call_tool_inner_disabled_tool_includes_help() {
-        use crate::mcp::permission::McpPermissionMode;
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::new().await;
-        let server = DataZenMcpServer::new(Arc::new(test.state))
-            .with_disabled_tools(&["query".into()])
-            .with_permission_mode(McpPermissionMode::SafeWrite);
-
-        let err = server
-            .call_tool_inner(
-                "query",
-                Some(serde_json::json!({"connection_id":"x","sql":"SELECT 1"})),
-            )
-            .await
-            .unwrap_err();
-        let msg = err.message.to_string();
-        assert!(msg.contains("disabled"));
-        assert!(msg.contains("Tool: query"));
-        assert!(msg.contains("connection_id"));
-    }
-
-    #[tokio::test]
-    async fn mcp_call_tool_inner_allowlist_error_includes_help() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        test.save_connection("blocked-cfg").await;
-        let server = DataZenMcpServer::new(Arc::new(test.state))
-            .with_allowed_connections(&["allowed-only".into()]);
-
-        let err = server
-            .call_tool_inner(
-                "list_databases",
-                Some(serde_json::json!({"connection_id":"blocked-cfg"})),
-            )
-            .await
-            .unwrap_err();
-        let msg = err.message.to_string();
-        assert!(msg.contains("allowlist"));
-        assert!(msg.contains("Tool: list_databases"));
-        assert!(msg.contains("Example:"));
-    }
-
-    #[tokio::test]
-    async fn mcp_query_rejects_disallowed_sql_in_readonly() {
-        use crate::mcp::permission::McpPermissionMode;
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        test.save_connection("ro-cfg").await;
-        let server = DataZenMcpServer::new(Arc::new(test.state))
-            .with_permission_mode(McpPermissionMode::ReadOnly);
-
-        let err = server
-            .query(rmcp::handler::server::wrapper::Parameters(QueryInput {
-                connection_id: "ro-cfg".into(),
-                sql: "DELETE FROM users".into(),
-                limit: None,
-            }))
-            .await
-            .unwrap_err();
-        assert!(
-            err.to_string().to_lowercase().contains("delete")
-                || err.to_string().contains("not allowed")
-                || err.to_string().contains("permission")
-        );
-    }
-
-    #[tokio::test]
-    async fn mcp_get_info_exposes_capabilities() {
-        use crate::testing::app_state::TestAppState;
-        use rmcp::ServerHandler;
-        use std::sync::Arc;
-
-        let test = TestAppState::new().await;
-        let server = DataZenMcpServer::new(Arc::new(test.state));
-        let info = server.get_info();
-        assert_eq!(info.server_info.name, "datazen");
-        assert!(info.capabilities.tools.is_some());
-    }
-
-    #[tokio::test]
-    async fn mcp_disabled_tools_and_permission_mode() {
-        use crate::mcp::permission::McpPermissionMode;
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::new().await;
-        let server = DataZenMcpServer::new(Arc::new(test.state))
-            .with_disabled_tools(&["query".into()])
-            .with_permission_mode(McpPermissionMode::ReadOnly);
-
-        let mut router = DataZenMcpServer::tool_router();
-        for name in MCP_ALL_TOOLS {
-            if !crate::mcp::permission::is_tool_listed(
-                name,
-                McpPermissionMode::ReadOnly,
-                &server.disabled_tools,
-            ) {
-                router.disable_route(name.to_string());
-            }
-        }
-        let listed = router.list_all();
-        assert!(!listed.iter().any(|t| t.name == "query"));
-        assert!(listed.iter().any(|t| t.name == "list_connections"));
-    }
-
-    #[tokio::test]
-    async fn mcp_list_tools_and_resources_inner() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::new().await;
-        let server = DataZenMcpServer::new(Arc::new(test.state));
-        assert_eq!(server.list_tools_inner().len(), MCP_ALL_TOOLS.len());
-        assert_eq!(server.list_resources_inner().len(), 3);
-    }
-
-    #[tokio::test]
-    async fn mcp_call_tool_inner_database_tools() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        test.save_connection("cti-cfg").await;
-        let server = DataZenMcpServer::new(Arc::new(test.state));
-
-        let dbs = server
-            .call_tool_inner(
-                "list_databases",
-                Some(serde_json::json!({"connection_id":"cti-cfg"})),
-            )
-            .await
-            .unwrap();
-        assert!(dbs.content[0].as_text().unwrap().text.contains("app"));
-
-        let tables = server
-            .call_tool_inner(
-                "list_tables",
-                Some(serde_json::json!({"connection_id":"cti-cfg","database":"app"})),
-            )
-            .await
-            .unwrap();
-        assert!(tables.content[0].as_text().unwrap().text.contains("users"));
-
-        let query = server
-            .call_tool_inner(
-                "query",
-                Some(serde_json::json!({"connection_id":"cti-cfg","sql":"SELECT 1","limit":5})),
-            )
-            .await
-            .unwrap();
-        assert!(!query.content[0].as_text().unwrap().text.is_empty());
-
-        assert!(server.call_tool_inner("missing_tool", None).await.is_err());
-    }
-
-    #[tokio::test]
-    async fn mcp_run_workflow_not_found() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::new().await;
-        let server = DataZenMcpServer::new(Arc::new(test.state));
-        let err = server
-            .run_workflow(rmcp::handler::server::wrapper::Parameters(
-                RunWorkflowInput {
-                    workflow_id: "missing".into(),
-                    variables: serde_json::json!({}),
-                    connection_id: None,
-                },
-            ))
-            .await
-            .unwrap_err();
-        assert!(err.to_string().contains("not found"));
-    }
-
-    #[tokio::test]
-    async fn mcp_query_history_read_only_returns_empty() {
-        use crate::mcp::permission::McpPermissionMode;
-        use crate::store::QueryHistoryEntry;
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        test.save_connection("hist-ro").await;
-        test.store
-            .add_query_history(QueryHistoryEntry {
-                id: "hist-ro-1".into(),
-                connection_id: "hist-ro".into(),
-                database: "app".into(),
-                schema: None,
-                sql: "SELECT secret FROM users".into(),
-                executed_at: chrono::Utc::now(),
-                execution_time_ms: 1,
-                rows_affected: None,
-                success: true,
-                error_message: None,
-            })
-            .await
-            .expect("add_query_history");
-
-        let server = DataZenMcpServer::new(Arc::new(test.state))
-            .with_permission_mode(McpPermissionMode::ReadOnly);
-
-        assert_eq!(server.list_resources_inner().len(), 2);
-        assert!(!server
-            .list_resources_inner()
-            .iter()
-            .any(|r| r.uri.as_str() == "datazen://query-history"));
-
-        let hist = server
-            .read_resource_inner("datazen://query-history")
-            .await
-            .unwrap();
-        match &hist.contents[0] {
-            ResourceContents::TextResourceContents { text, .. } => {
-                assert_eq!(text.trim(), "[]");
-                assert!(!text.contains("secret"));
-            }
-            _ => panic!("expected text resource"),
-        }
-    }
-
-    #[tokio::test]
-    async fn mcp_query_history_respects_connection_allowlist() {
-        use crate::store::QueryHistoryEntry;
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        test.save_connection("hist-allowed").await;
-        test.save_connection("hist-blocked").await;
-        for (id, sql) in [("hist-a", "SELECT 1"), ("hist-b", "SELECT 2")] {
-            test.store
-                .add_query_history(QueryHistoryEntry {
-                    id: id.into(),
-                    connection_id: if id == "hist-a" {
-                        "hist-allowed".into()
-                    } else {
-                        "hist-blocked".into()
-                    },
-                    database: "app".into(),
-                    schema: None,
-                    sql: sql.into(),
-                    executed_at: chrono::Utc::now(),
-                    execution_time_ms: 1,
-                    rows_affected: None,
-                    success: true,
-                    error_message: None,
-                })
-                .await
-                .expect("add_query_history");
-        }
-
-        let server = DataZenMcpServer::new(Arc::new(test.state))
-            .with_allowed_connections(&["hist-allowed".into()]);
-
-        let hist = server
-            .read_resource_inner("datazen://query-history")
-            .await
-            .unwrap();
-        match &hist.contents[0] {
-            ResourceContents::TextResourceContents { text, .. } => {
-                assert!(text.contains("hist-allowed") || text.contains("\"connectionId\""));
-                assert!(text.contains("SELECT 1"));
-                assert!(!text.contains("hist-blocked"));
-                assert!(!text.contains("SELECT 2"));
-            }
-            _ => panic!("expected text resource"),
-        }
-    }
-
-    #[tokio::test]
-    async fn mcp_connections_resource_respects_allowlist() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        test.save_connection("conn-visible").await;
-        test.save_connection("conn-hidden").await;
-
-        let server = DataZenMcpServer::new(Arc::new(test.state))
-            .with_allowed_connections(&["conn-visible".into()]);
-
-        let conns = server
-            .read_resource_inner("datazen://connections")
-            .await
-            .unwrap();
-        match &conns.contents[0] {
-            ResourceContents::TextResourceContents { text, .. } => {
-                assert!(text.contains("conn-visible"));
-                assert!(!text.contains("conn-hidden"));
-            }
-            _ => panic!("expected text resource"),
-        }
-    }
-
-    #[tokio::test]
-    async fn mcp_schema_resource_rejects_disallowed_connection() {
-        use crate::testing::app_state::TestAppState;
-        use std::sync::Arc;
-
-        let test = TestAppState::with_tables().await;
-        test.save_connection("schema-blocked").await;
-        let server = DataZenMcpServer::new(Arc::new(test.state))
-            .with_allowed_connections(&["other-only".into()]);
-
-        let err = server
-            .read_resource_inner("datazen://schema/schema-blocked/app")
-            .await
-            .unwrap_err();
-        assert!(err.to_string().contains("allowlist"));
-    }
+    // Remaining integration tests live in server.rs; include from extracted block below.
+    include!("server/tests_integration.rs");
 }
